requirements.txt
langgraph>=0.2.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
pydantic>=2.5.0
python-dotenv>=1.0.0
pypdf>=4.0.0
typing-extensions>=4.8.0

